update job result type
