+++ qiskit_ibm_runtime/estimator.py
@@ -16,7 +16,6 @@ from __future__ import annotations
 import copy
 from typing import Iterable, Optional, Dict, Sequence, Any, Union
 
-import numpy as np
 from qiskit.circuit import QuantumCircuit, Parameter
 from qiskit.quantum_info import SparsePauliOp
 from qiskit.opflow import PauliSumOp
@@ -28,7 +27,7 @@ import qiskit_ibm_runtime.session as session_pkg
 # TODO import BaseEstimator and EstimatorResult from terra once released
 from .qiskit.primitives import BaseEstimator, EstimatorResult
 from .qiskit_runtime_service import QiskitRuntimeService
-from .program.result_decoder import ResultDecoder
+from .estimator_result_decoder import EstimatorResultDecoder
 from .runtime_job import RuntimeJob
 from .utils.deprecation import deprecate_arguments, issue_deprecation_msg
 from .runtime_options import RuntimeOptions
@@ -320,16 +319,3 @@ class Estimator(BaseEstimator):
             Session used by this primitive.
         """
         return self._session
-
-
-class EstimatorResultDecoder(ResultDecoder):
-    """Class used to decode estimator results"""
-
-    @classmethod
-    def decode(cls, raw_result: str) -> EstimatorResult:
-        """Convert the result to EstimatorResult."""
-        decoded: Dict = super().decode(raw_result)
-        return EstimatorResult(
-            values=np.asarray(decoded["values"]),
-            metadata=decoded["metadata"],
-        )
