+++ qiskit_ibm_runtime/runtime_job.py
@@ -23,6 +23,8 @@ from datetime import datetime
 
 from qiskit.providers.backend import Backend
 from qiskit.providers.jobstatus import JobStatus, JOB_FINAL_STATES
+from .estimator_result_decoder import EstimatorResultDecoder
+from .sampler_result_decoder import SamplerResultDecoder
 
 from .constants import API_TO_JOB_ERROR_MESSAGE, API_TO_JOB_STATUS
 from .exceptions import (
@@ -177,6 +179,10 @@ class RuntimeJob:
             RuntimeJobFailureError: If the job failed.
             RuntimeJobMaxTimeoutError: If the job does not complete within given timeout.
         """
+        if self.program_id == "sampler":
+            self._result_decoder = SamplerResultDecoder
+        if self.program_id == "estimator":
+            self._result_decoder = EstimatorResultDecoder
         _decoder = decoder or self._result_decoder
         if self._results is None or (_decoder != self._result_decoder):
             self.wait_for_final_state(timeout=timeout)
